No more app global in RemoteServer, move comment parsing to new file.
